refactor test.package : 패키지 종류별로 분류
