+++ src/test/java/gift/study/SerializeTest.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.study;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
