+++ src/test/java/gift/wish/WishServiceTest.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.wish;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
